[DEV-2296] refactor api section component (#1451)
* Refactor ApiSection into a generic section and rename the previous module to ApiRestSection

* Add changeset

* Update .changeset/purple-clocks-sniff.md


* Update apps/nextjs-website/src/components/molecules/ApiSection/ApiSection.tsx


* Move ApiRestSection to its own directory and rename its style.ts

---------
