+++ .changeset/purple-clocks-sniff.md
@@ -1,0 +1,5 @@
+---
+"nextjs-website": minor
+---
+
+Rename the existing ApiSection component to ApiRestSection and add a more generic ApiSection component to replace it
