+++ apps/nextjs-website/src/app/[productSlug]/api/[apiDataSlug]/page.tsx
@@ -3,7 +3,6 @@ import ProductLayout, {
   ProductLayoutProps,
 } from '@/components/organisms/ProductLayout/ProductLayout';
 import { Product } from '@/lib/types/product';
-import ApiSection from '@/components/molecules/ApiSection/ApiSection';
 import { Metadata, ResolvingMetadata } from 'next';
 import {
   makeMetadata,
@@ -18,6 +17,7 @@ import {
   convertApiToStructuredDataSoftwareApplication,
   productToBreadcrumb,
 } from '@/helpers/structuredData.helpers';
+import ApiSection from '@/components/molecules/ApiSection/ApiSection';
 
 export type ApiDataPageProps = {
   readonly title?: string;
@@ -98,12 +98,7 @@ const ApiDataPage = async ({ params }: ApiDataParams) => {
         showBreadcrumbs
         structuredData={structuredData}
       >
-        <ApiSection
-          apiSlug={params.apiDataSlug}
-          specURLs={apiDataProps.specURLs}
-          product={apiDataProps.product}
-          specURLsName={apiDataProps.specURLsName}
-        />
+        <ApiSection apiData={apiDataProps} />
       </ProductLayout>
     );
   }
